+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Vse pomembne spremembe tega projekta bodo dokumentirane v tej datoteki.
 
 Format je v skladu s [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [CURRENT](https://github.com/tomaz1/ffmpeg_convert) - 2025-06-16
+### Changed
+- When converting from `.mkv` to `.mp4` and if `.mkv` has subtitle/s it is skipped since `.mp4` supports only mov_text type and we would need to convert (possible TODO)
+- Better support for max video bitrate setting, to find bitrate of a video stream.
+
 ## [v1.7](https://github.com/tomaz1/ffmpeg_convert/releases/tag/v1.7) - 2025-06-16
 ### Added
 - Added support for `--max-video-bitrate N` option to force video conversion if bitrate exceeds N kbps.